django-scim/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
Django==5.1
django-extensions==3.2.3
django-oauth-toolkit==2.4.0
django-scim2==0.19.1
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
pycparser==2.22
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
scim2-filter-parser==0.7.0
simplejson==3.19.3
sly==0.5
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
